core: Cover MMU implementation into C++
This commit wraps the existing MMU implementation into C++ class. Now
MMU is an object which is dynamically created during runtime. To create
memory mapping the user needs just to create MMap object with respective
parameters. Destroying of this object will destroy the mapping.

The existing MMU implementation contains lots of TBDs, so refactoring
and follow up development is needed.

Signed-off-by: Alexander Smirnov <alex.bluesman.smirnov@gmail.com>