main: Add marker for safe stack reset
When we switch from EL2 to EL1 we don't need Saturn stack anymore and
just to avoid memory leak - it makes sense to reset stack pointer. But
the stack could not be fully reset, because there is one Saturn object
allocated on stack - Heap class. So this patch introduces a marker to
which SP could be safely reset.

Signed-off-by: Alexander Smirnov <alex.bluesman.smirnov@gmail.com>