console: Introduce blocking and non-blocking mode
For getting chars from input pipeline we need 2 modes: waiting and
non-waiting. For example when console application expects input from
user - it would be very comfortable to use blocking mode to avoid
polling mechanism implementation. But when VM tries to read UART RX
register - we should not block the system and should return to EL1
context as soon as possible even if input pipeline is empty (just return
0).

Signed-off-by: Alexander Smirnov <alex.bluesman.smirnov@gmail.com>